优化外接设备处理

+++ src/com/cantv/media/center/activity/GridViewActivity.java
@@ -1,6 +1,8 @@
 package com.cantv.media.center.activity;
+
 import java.util.ArrayList;
 import java.util.List;
+
 import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
@@ -12,6 +14,7 @@ import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
+
 import com.cantv.media.R;
 import com.cantv.media.center.constants.SourceType;
 import com.cantv.media.center.data.Media;
@@ -47,7 +50,8 @@ public class GridViewActivity extends Activity {
     private List<MenuItem> viewModeSubMenuItems;
     private List<MenuItem> mMenuList;
     private int mSelectedMenuPosi;
-    private int mDeleteItem ;
+    private int mDeleteItem;
+
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -56,6 +60,9 @@ public class GridViewActivity extends Activity {
         mContentView = (RelativeLayout) findViewById(R.id.gridview_content);
         Intent intent = getIntent();
         Uri uri = intent.getData();
+
+        String deviceFlag = getIntent().getStringExtra("toListFlag");
+
         String type = null;
         if (uri != null) {
             type = uri.getHost();
@@ -84,61 +91,67 @@ public class GridViewActivity extends Activity {
                     SourceType.LOCAL);
         } else if ("device1".equalsIgnoreCase(type)) {
             mTitleTV.setText(R.string.str_external);
-            String filePath = getIntent().getStringExtra("filePath");
+            String filePath = "";
+            if ((null == deviceFlag) && MediaUtils.isExistUSB()) {
+                filePath = MediaUtils.getUsbRootPaths().get(0);
+            }
+
             mGridView = new MediaGridView(this, filePath, SourceType.DEVICE);
-           
+
         } else if ("device2".equalsIgnoreCase(type)) {
             mTitleTV.setText(R.string.str_external);
             String filePath = getIntent().getStringExtra("filePath");
             mGridView = new MediaGridView(this, filePath, SourceType.DEVICE);
         }
-        String deviceFlag = getIntent().getStringExtra("toListFlag");
-        if(null!=deviceFlag){
-        	mGridView.mDevecesFlag=deviceFlag;
+        if (null != deviceFlag) {
+            mGridView.mDevecesFlag = deviceFlag;
         }
         mGridView.show();
         mContentView.removeAllViews();
         switch (SharedPreferenceUtil.getGridStyle()) {
-		case 0:
-			setGridStyle(MediaOrientation.THUMBNAIL);
-			break;
-		case 1:
-			setGridStyle(MediaOrientation.LIST);
-			break;
-		}
+            case 0:
+                setGridStyle(MediaOrientation.THUMBNAIL);
+                break;
+            case 1:
+                setGridStyle(MediaOrientation.LIST);
+                break;
+        }
         mContentView.addView(mGridView);
     }
+
     @Override
     public boolean onKeyDown(int keyCode, KeyEvent event) {
         if (8 == keyCode || 166 == keyCode) {
             setGridStyle(MediaOrientation.THUMBNAIL);
             SharedPreferenceUtil.setGridStyle(0);
         } else if (9 == keyCode || 167 == keyCode) {
-        	SharedPreferenceUtil.setGridStyle(1);
+            SharedPreferenceUtil.setGridStyle(1);
             setGridStyle(MediaOrientation.LIST);
         } else if (keyCode == KeyEvent.KEYCODE_MENU) {
             showMenuDialog();
         }
         return super.onKeyDown(keyCode, event);
     }
+
     public void setGridStyle(MediaOrientation mStyle) {
         switch (mStyle) {
-        case LIST:
-            mGridView.setVerticalSpacing((int) getResources().getDimension(
-                    R.dimen.px31));
-            mGridView.setPadding(0, 0, 0, 0);
-            mGridView.setStyle(MediaOrientation.LIST);
-            mGridView.setNumColumns(1);
-            break;
-        case THUMBNAIL:
-            mGridView.setVerticalSpacing((int) getResources().getDimension(
-                    R.dimen.px0));
-            mGridView.setPadding(0, 0, 0, 22);
-            mGridView.setStyle(MediaOrientation.THUMBNAIL);
-            mGridView.setNumColumns(5);
-            break;
+            case LIST:
+                mGridView.setVerticalSpacing((int) getResources().getDimension(
+                        R.dimen.px31));
+                mGridView.setPadding(0, 0, 0, 0);
+                mGridView.setStyle(MediaOrientation.LIST);
+                mGridView.setNumColumns(1);
+                break;
+            case THUMBNAIL:
+                mGridView.setVerticalSpacing((int) getResources().getDimension(
+                        R.dimen.px0));
+                mGridView.setPadding(0, 0, 0, 22);
+                mGridView.setStyle(MediaOrientation.THUMBNAIL);
+                mGridView.setNumColumns(5);
+                break;
         }
     }
+
     @Override
     public void onBackPressed() {
         // MediaGridView childGridView = (MediaGridView)
@@ -151,6 +164,7 @@ public class GridViewActivity extends Activity {
             return;
         }
     }
+
     private void showMenuDialog() {
         if (mMenuDialog == null) {
             mMenuDialog = new MenuDialog(this);
@@ -159,14 +173,15 @@ public class GridViewActivity extends Activity {
             mMenuDialog.setOnItemClickListener(new OnItemClickListener() {
                 @Override
                 public void onSubMenuItemClick(LinearLayout parent, View view,
-                        int position) {
+                                               int position) {
                     subMenuClick(position);
                 }
+
                 @Override
                 public boolean onMenuItemClick(LinearLayout parent, View view,
-                        int position) {
-                    if(position != 2){
-                        if(mSelectedMenuPosi == position){
+                                               int position) {
+                    if (position != 2) {
+                        if (mSelectedMenuPosi == position) {
                             return false;
                         }
                     }
@@ -175,31 +190,37 @@ public class GridViewActivity extends Activity {
                     MenuItem menuItem = mMenuList.get(position);
                     menuItem.setSelected(true);
                     mMenuDialog.getMenuAdapter().notifySubMenuDataSetChanged();
-                    if(position == 2){
-                        mDeleteItem = mGridView.mSelectItemPosition ; 
+                    if (position == 2) {
+                        mDeleteItem = mGridView.mSelectItemPosition;
                         List<Media> datas = mGridView.mListAdapter.getData();
-                        Media media = datas.get(mDeleteItem);
-                        boolean deleteSuccessed = FileUtil.delete(media);
-                        if (deleteSuccessed) {
-                            datas.remove(mDeleteItem);
-                            mGridView.mListAdapter.bindData(datas);
-                            Log.i("shen", "mDeleteItem:" + mDeleteItem);
+
+                        if (datas.size() > 0) {  //防止当前目录没有数据,进行删除操作发生异常
+                            Media media = datas.get(mDeleteItem);
+                            boolean deleteSuccessed = FileUtil.delete(media);
+                            if (deleteSuccessed) {
+                                datas.remove(mDeleteItem);
+                                mGridView.mListAdapter.bindData(datas);
+                                Log.i("shen", "mDeleteItem:" + mDeleteItem);
+                            } else {
+                                Toast.makeText(GridViewActivity.this, R.string.deleteFailed, Toast.LENGTH_SHORT).show();
+                            }
                         } else {
-                            Toast.makeText(GridViewActivity.this, R.string.deleteFailed, Toast.LENGTH_SHORT).show();
+                            Toast.makeText(GridViewActivity.this, "没有数据!", Toast.LENGTH_SHORT).show();
                         }
-                        
-                        return true ;
-                    }else{
-                    	return false ;
+
+                        return true;
+                    } else {
+                        return false;
                     }
                 }
             });
         }
         mMenuDialog.show();
     }
+
     private List<MenuItem> createMenuData() {
         mMenuList = new ArrayList<MenuItem>();
-        
+
         sortListMenuItem = new MenuItem(getString(R.string.sort));
         sortListMenuItem.setType(MenuItem.TYPE_SELECTOR);
         sortListMenuItem.setSelected(true);
@@ -213,7 +234,7 @@ public class GridViewActivity extends Activity {
         sortMenu.setSelected(true);
         sortListMenuItem.setChildren(sortListSubMenuItems);
         mMenuList.add(sortListMenuItem);
-        
+
         viewModeMenuItem = new MenuItem(getString(R.string.view));
         viewModeMenuItem.setType(MenuItem.TYPE_SELECTOR);
         viewModeSubMenuItems = new ArrayList<MenuItem>();
@@ -225,13 +246,14 @@ public class GridViewActivity extends Activity {
         viewMenu.setSelected(true);
         viewModeMenuItem.setChildren(viewModeSubMenuItems);
         mMenuList.add(viewModeMenuItem);
-        
+
         deleteMenuItem = new MenuItem(getString(R.string.delete));
         deleteMenuItem.setType(MenuItem.TYPE_NORMAL);
         mMenuList.add(deleteMenuItem);
-        
+
         return mMenuList;
     }
+
     private void subMenuClick(int position) {
         boolean isRefreshed = false;
         MenuItem menuItemData = mMenuList.get(mSelectedMenuPosi);
@@ -259,18 +281,18 @@ public class GridViewActivity extends Activity {
             } else if (position == 1) {
                 setGridStyle(MediaOrientation.THUMBNAIL);
             }
-        } 
-        
+        }
+
         View oldSubMenuItemView = mMenuDialog.getMenu().findViewWithTag(MenuAdapter.TAG_SUB_MENU_VIEW + lastSelectPosi);
-        if(oldSubMenuItemView != null){
+        if (oldSubMenuItemView != null) {
             mMenuDialog.getMenuAdapter().updateSubMenuItem(oldSubMenuItemView, menuItemData.getChildAt(lastSelectPosi));
         }
         View subMenuItemView = mMenuDialog.getMenu().findViewWithTag(MenuAdapter.TAG_SUB_MENU_VIEW + position);
-        if(subMenuItemView != null){
+        if (subMenuItemView != null) {
             mMenuDialog.getMenuAdapter().updateSubMenuItem(subMenuItemView, menuItemData.getSelectedChild());
         }
         View menuItemView = mMenuDialog.getMenu().findViewWithTag(MenuAdapter.TAG_MENU_VIEW + mSelectedMenuPosi);
-        if(menuItemView != null){
+        if (menuItemView != null) {
             mMenuDialog.getMenuAdapter().updateMenuItem(menuItemView, menuItemData);
         }
     }
